Fixed some compiler warnings on several classes Added javadoc on Query Added a license and javadoc to QueryFrame
git-svn-id: http://svn.dmdirc.com/trunk@273 00569f92-eb28-0410-84fd-f71c24880f43

+++ src/uk/org/ownage/dmdirc/parser/ChannelClientInfo.java
@@ -95,7 +95,7 @@ public class ChannelClientInfo {
 				// There may be a better alternative to this, but this works well. (Gets the prefix modes in order)
 				for (Enumeration e = myParser.hPrefixModes.keys(); e.hasMoreElements();) {
 					cTemp = (Character)e.nextElement();
-					nTemp = (Integer)myParser.hPrefixModes.get(cTemp);
+					nTemp = myParser.hPrefixModes.get(cTemp);
 					if (nTemp == i) {
 						if (bPrefix) { cTemp = myParser.hPrefixMap.get(cTemp); }
 						sModes = sModes+cTemp;
